requirements.txt
appium-python-client>=3.0.0
pytest>=7.0
pytest-html>=3.2
webdriver-manager
loguru